# contract/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Meeting Slot Booking API
  description: Draft API contract. Organizer management operations use a per-organizer bearer management key. Public booking operations do not require it.
  version: 0.0.0
...
paths:
...
  /v1/organizers:
    post:
      operationId: createOrganizer
      description: Creates an empty organizer workspace. Anyone may create one. The response includes the organizer ID and management key, and the service also emails both to organizerEmail for later access. There is no key recovery endpoint. organizerEmail receives booking confirmation and cancellation notifications. Temporary hold credentials and SMTP configuration credentials are never emailed.
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerCreatedResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRequest'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizerRequest'
  /v1/organizers/me:
    get:
      operationId: getOrganizerIdentity
      description: Resolves the organizer ID associated with the bearer management key.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerIdentityResponse'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthenticated'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
      security:
        - BearerAuth: []
  /v1/organizers/{organizerId}/bookings:
    get:
      operationId: listOrganizerBookings
      description: Requires the bearer management key scoped to this organizer ID.
      parameters:
        - name: organizerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerBookingsResponse'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthenticated'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerAccessDenied'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
      security:
        - BearerAuth: []
  /v1/organizers/{organizerId}/bookings/{bookingId}/cancellation:
    post:
      operationId: cancelBookingAsOrganizer
      description: Cancels a confirmed booking, subject to the 24-hour cutoff. Requires the bearer management key scoped to this organizer ID.
      parameters:
        - name: organizerId
          in: path
          required: true
          schema:
            type: string
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerCancellationResponse'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthenticated'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerAccessDenied'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateConflict'
        '422':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancellationTooLate'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
      security:
        - BearerAuth: []
  /v1/organizers/{organizerId}/slots:
    post:
      operationId: createOrganizerSlot
      description: Requires the bearer management key scoped to this organizer ID.
      parameters:
        - name: organizerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRequest'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthenticated'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerAccessDenied'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/StateConflict'
                  - $ref: '#/components/schemas/SlotConflict'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSlotRequest'
      security:
        - BearerAuth: []
    get:
      operationId: listOrganizerSlots
      description: Requires the bearer management key scoped to this organizer ID.
      parameters:
        - name: organizerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotsResponse'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthenticated'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerAccessDenied'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
      security:
        - BearerAuth: []
  /v1/organizers/{organizerId}/slots/{slotId}:
    delete:
      operationId: removeOrganizerSlot
      description: Removes an unbooked slot. Requires the bearer management key scoped to this organizer ID.
      parameters:
        - name: organizerId
          in: path
          required: true
          schema:
            type: string
        - name: slotId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthenticated'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerAccessDenied'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateConflict'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
      security:
        - BearerAuth: []
  /v1/public/bookings/cancellation:
    post:
      operationId: cancelBookingWithCredential
      description: Cancels the booking identified by the unique X-Cancellation-Credential. Missing and unmatched credentials both return CANCELLATION_CREDENTIAL_INVALID.
      parameters:
        - $ref: '#/components/parameters/ClientCancellationRequest'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancellationResponse'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthenticated'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateConflict'
        '422':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancellationTooLate'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
  /v1/public/bookings/{bookingId}/cancellation:
    post:
      operationId: cancelBookingAsClient
      description: Legacy compatibility operation. Cancels a booking by ID after validating X-Cancellation-Credential.
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/ClientCancellationRequest'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancellationResponse'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthenticated'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateConflict'
        '422':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancellationTooLate'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
  /v1/public/holds/{holdId}/confirmation:
    post:
      operationId: confirmHold
      description: Confirms a booking and subscribes clientEmail to updates for this booking. The response includes the booking ID and cancellation credential, and the service also emails both to clientEmail for later access. The organizer and client receive confirmation and later cancellation notifications; if both addresses match, the client email includes the cancellation details. Temporary hold credentials and SMTP configuration credentials are never emailed.
      parameters:
        - name: holdId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRequest'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthenticated'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateConflict'
        '410':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HoldExpired'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmBookingRequest'
  /v1/public/organizers/{organizerId}/availability:
    get:
      operationId: getPublicAvailability
      description: Returns only currently available slots. Held, booked, and withdrawn slots are excluded. When a hold expires, its slot becomes eligible for availability again.
      parameters:
        - name: organizerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
  /v1/public/organizers/{organizerId}/holds:
    post:
      operationId: createHold
      parameters:
        - name: organizerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HoldResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRequest'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateConflict'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHoldRequest'
components:
...
  schemas:
...
    CreateOrganizerRequest:
      type: object
      required:
        - organizerEmail
      properties:
        organizerEmail:
          $ref: '#/components/schemas/ClientEmail'
          description: Email address that receives notifications about this organizer's bookings.
...
    InternalError:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - INTERNAL_ERROR
              description: Stable machine-readable error identifier; the allowed value depends on the response status.
            message:
              type: string
              description: Human-readable explanation safe to show to the caller.
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
              description: Optional field-level validation details.
            requestId:
              type: string
              description: Request correlation identifier when available.
    InvalidRequest:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
              description: Stable machine-readable error identifier; the allowed value depends on the response status.
            message:
              type: string
              description: Human-readable explanation safe to show to the caller.
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
              description: Optional field-level validation details.
            requestId:
              type: string
              description: Request correlation identifier when available.
...
    OrganizerBookingsResponse:
      type: object
      required:
        - bookings
      properties:
        bookings:
          type: array
          items:
            $ref: '#/components/schemas/OrganizerBookingDetails'
    OrganizerCancellationResponse:
      type: object
      required:
        - booking
      properties:
        booking:
          $ref: '#/components/schemas/OrganizerBookingDetails'
    OrganizerCreatedResponse:
      type: object
      required:
        - organizerId
        - managementKey
      properties:
        organizerId:
          type: string
        managementKey:
          type: string
          description: Cryptographically random 256-bit management key; returned when this organizer is created and also emailed with the organizer ID.
    OrganizerIdentityResponse:
      type: object
      required:
        - organizerId
      properties:
        organizerId:
          type: string
...
    SlotResponse:
      type: object
      required:
        - slot
      properties:
        slot:
          $ref: '#/components/schemas/Slot'
    SlotsResponse:
      type: object
      required:
        - slots
      properties:
        slots:
          type: array
          items:
            $ref: '#/components/schemas/OrganizerSlot'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: Bearer
servers:
  - url: https://api.example.com
    description: Draft API server
    variables: {}